3_largest_prime_factor.cpp

+++ 3_largest_prime_factor.cpp
@@ -6,9 +6,9 @@
 constexpr ll NUM = 600'851'475'143;
 using namespace std;
 
-vector<int> get_factors(ll val) {
-    vector<int> factor;
-    for (int i = 2; i * i <= val; ++i) {
+vector<ll> get_factors(ll val) {
+    vector<ll> factor;
+    for (ll i = 2; i * i <= val; ++i) {
         if (val % i == 0) {
             factor.push_back(i);
         }
@@ -20,7 +20,7 @@ vector<int> get_factors(ll val) {
 int main() {
     const auto factors = get_factors(NUM);
 
-    for (int i = static_cast<int>(factors.size()) - 1; i >= 0; --i) {
+    for (ll i = static_cast<int>(factors.size()) - 1; i >= 0; --i) {
         if (auto val = get_factors(factors[i]); val.empty()) {
             cout << i;
